Add comments
So I understand what the code does without needing to read it
thoroughly. This probably means I should make the logic a bit easier to
read and digest, but I'm happy with having something working right now.

The comments will help a great deal if I come back to this and
rename/add functions to improve the readability.